Started work on module 4
Added the bare minimum for the other pages on the website, but did properly implement the HTML outline for wizard.html.
Fixed an oversight in the styles.css which allowed text to escape a <section> if zoomed in too much. Also generally cleaned up how <section> logic works to be generally pleasable. 
Added new images for usage on wizard.html, and replaced the logo with a more proper one.