Polished resizing logic.
Changed text sizing logic from based on viewport to EM size. This has the drawback of not being able to scale to 4K monitors with horrible DPI, but even GitHub website doesn't do that! Worth it for better and more traditional resizing logic.
Changed image scaling logic to be more forgiving. More tall, and is only sticky if image would not be next to text at all.
Made E-Card render canvas scale to resizing of page.